requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
docker==5.0.3
Flask==2.0.2
Flask-Compress==1.10.1
Flask-JWT-Extended==4.3.1
flask-restx==0.5.1
idna==3.3
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydotenv==0.0.7
PyJWT==2.3.0
pymongo==4.0
pyparsing==3.0.7
pyrsistent==0.18.0
pytest==7.0.1
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.7
waitress==2.0.0
websocket-client==1.2.3
Werkzeug==2.0.2
zipp==3.6.0
